refactor(frontend): unify quick-reaction recents onto per-server emoji store
The message hover bar, context menu, and mobile action sheet now fill their
non-pinned slots from the same per-server `recentEmojis` store the picker
uses. Replaces the old global `recentReactions` store (which kept its own
non-namespaced list); switching servers now shows distinct recents
everywhere, and emojis picked from the full picker immediately appear in
the quick-reaction slots on the same server.

- Add `quickReactions` getter to `RecentEmojisStore` (pinned + non-pinned
  recents + fallbacks, capped at QUICK_REACTIONS_COUNT)
- MessageHoverBar/MessageContextMenu/MessageActionSheet take a `serverId`
  prop and read from `getRecentEmojis(serverId).quickReactions`
- Delete obsolete `recentReactions` store and spec
